+++ ubelt/util_path.py
@@ -410,6 +410,7 @@ class Path(_PathBase):
         ~/.cache/ubelt/demo_path/text_file.aux.jpg
         ~/.cache/ubelt/demo_pathdemo_path2
     """
+    __slots__ = ()
 
     def touch(self, mode=0o666, exist_ok=True):
         """
@@ -638,3 +639,74 @@ class Path(_PathBase):
             return cls(util_platform.get_app_data_dir(appname, *args))
         else:
             raise KeyError(type)
+
+    def ls(self):
+        """
+        A convenience function to list all paths in a directory.
+
+        This is simply a wraper around iterdir that returns the results as a
+        list instead of a generator. This is mainly for faster navigation in
+        IPython. In production code `iterdir` should be used instead.
+
+        Returns:
+            List[Path]
+
+        Example:
+            >>> import ubelt as ub
+            >>> self = ub.Path.appdir('ubelt/tests/ls')
+            >>> (self / 'dir1').ensuredir()
+            >>> (self / 'dir2').ensuredir()
+            >>> (self / 'file1').touch()
+            >>> (self / 'file2').touch()
+            >>> (self / 'dir1/file3').touch()
+            >>> (self / 'dir2/file4').touch()
+            >>> children = self.ls()
+            >>> assert isinstance(children, list)
+            >>> print(ub.repr2(sorted([p.relative_to(self) for p in children])))
+            [
+                Path('dir1'),
+                Path('dir2'),
+                Path('file1'),
+                Path('file2'),
+            ]
+        """
+        return list(self.iterdir())
+
+    def walk(self, topdown=True, onerror=None, followlinks=False):
+        """
+        A variant of os.walk for pathlib
+
+        Args:
+            topdown (bool):
+                if True starts yield nodes closer to the root first otherwise
+                yield nodes closer to the leaves first.
+
+            onerror (callable):
+                A function with one argument of type OSError. If the
+                error is raised the walk is aborted, otherwise it continues.
+
+            followlinks (bool):
+                if True recurse into symbolic directory links
+
+        Yields:
+            Tuple[Path, str, str]: the root, file names, and directory names
+
+        Example:
+            >>> import ubelt as ub
+            >>> self = ub.Path.appdir('ubelt/tests/ls')
+            >>> (self / 'dir1').ensuredir()
+            >>> (self / 'dir2').ensuredir()
+            >>> (self / 'file1').touch()
+            >>> (self / 'file2').touch()
+            >>> (self / 'dir1/file3').touch()
+            >>> (self / 'dir2/file4').touch()
+            >>> subdirs = list(self.walk())
+            >>> assert len(subdirs) == 3
+        """
+        import os
+        cls = self.__class__
+        walker = os.walk(self, topdown=topdown, onerror=onerror,
+                         followlinks=followlinks)
+        for r, fs, ds in walker:
+            r = cls(r)
+            yield r, fs, ds